Refactor arq/arq7.rs into smaller files
