+++ arq/src/arq7/backup_folder.rs
@@ -1,0 +1,71 @@
+use super::backup_folders::{is_file_encrypted, decrypt_json_file};
+use super::encrypted_keyset::EncryptedKeySet;
+use crate::error::Result;
+use serde::{Deserialize, Serialize};
+use std::fs::File;
+use std::io::BufReader;
+use std::path::Path;
+
+/// BackupFolder represents a backupfolder.json file within the backupfolders/<UUID>/ directory
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct BackupFolder {
+    #[serde(rename = "localPath")]
+    pub local_path: String,
+    #[serde(rename = "migratedFromArq60")]
+    pub migrated_from_arq60: bool,
+    #[serde(rename = "storageClass")]
+    pub storage_class: String,
+    #[serde(default)]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "diskIdentifier")]
+    pub disk_identifier: Option<String>,
+    pub uuid: String,
+    #[serde(rename = "migratedFromArq5")]
+    pub migrated_from_arq5: bool,
+    #[serde(rename = "localMountPoint")]
+    pub local_mount_point: String,
+    pub name: String,
+}
+
+impl BackupFolder {
+    /// Load a BackupFolder from a JSON reader
+    pub fn from_reader<R: std::io::Read>(reader: R) -> Result<Self> {
+        Ok(serde_json::from_reader(reader)?)
+    }
+
+    /// Load a BackupFolder from a file path
+    pub fn from_file<P: AsRef<std::path::Path>>(path: P) -> Result<Self> {
+        Self::from_file_with_encryption(path, None)
+    }
+
+    /// Load BackupFolder from file, optionally decrypting if needed
+    pub fn from_file_with_encryption<P: AsRef<Path>>(
+        path: P,
+        keyset: Option<&EncryptedKeySet>,
+    ) -> Result<BackupFolder> {
+        load_json_with_encryption(path, keyset)
+    }
+}
+
+/// Helper function to load JSON from either encrypted or unencrypted file
+fn load_json_with_encryption<T, P>(path: P, keyset: Option<&EncryptedKeySet>) -> Result<T>
+where
+    T: for<'de> serde::Deserialize<'de>,
+    P: AsRef<Path>,
+{
+    let path_ref = path.as_ref();
+
+    if let Some(keyset) = keyset {
+        // Check if file is encrypted
+        if is_file_encrypted(path_ref)? {
+            // Decrypt and parse
+            let json_content = decrypt_json_file(path_ref, keyset)?;
+            return Ok(serde_json::from_str(&json_content)?);
+        }
+    }
+
+    // Load as regular unencrypted file
+    let file = File::open(path_ref)?;
+    let reader = BufReader::new(file);
+    return Ok(serde_json::from_reader(reader)?);
+}
